+++ shims/spark40/src/main/java/org/apache/spark/sql/vectorized/ArrowColumnarArray.java
@@ -14,23 +14,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gluten.vectorized;
+package org.apache.spark.sql.vectorized;
 
-import org.apache.spark.sql.execution.vectorized.ColumnarArrayShim;
-import org.apache.spark.sql.vectorized.ColumnVector;
+import org.apache.spark.unsafe.types.VariantVal;
 
-/**
- * Because `get` method in `ColumnarArray` don't check whether the data to get is null and arrow
- * vectors will throw exception when we try to access null value, so we define the following class
- * as a workaround. Its implementation is copied from Spark-4.0, except that the `handleNull`
- * parameter is set to true when we call `SpecializedGettersReader.read` in `get`, which means that
- * when trying to access a value of the array, we will check whether the value to get is null first.
- *
- * <p>The actual implementation is put in [[ColumnarArrayShim]] because Variant data type is
- * introduced in Spark-4.0.
- */
-public class ArrowColumnarArray extends ColumnarArrayShim {
+public final class ArrowColumnarArray extends AbstractColumnarArray {
+
   public ArrowColumnarArray(ColumnVector data, int offset, int length) {
     super(data, offset, length);
+  }
+
+  @Override
+  public VariantVal getVariant(int ordinal) {
+    return data.getVariant(offset + ordinal);
   }
 }
